requirements.txt
langchain-openai
python-dotenv
streamlit
pandas 